pyproject.toml
[project]
name = "pinn3d"
version = "0.1.0"
description = "3D PINN for Helmholtz equation on unit cube"
requires-python = ">=3.12"
dependencies = [
    "jax==0.9.0",
    "equinox==0.13.2",
    "fastapi==0.124.0",
    "pyngrok==7.5.0",
    "uvicorn",
    "optax",
    "numpy",
    "scipy",
    "pydantic",
    "matplotlib",
    "pyyaml",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
